Modify to classify data to be stored in DB and data to be directly extracted
:Release Notes:
Modify to classify data to be stored in DB
and data to be directly extracted, and fix
bugs

:Detailed Notes:
1. To reduce the number of meta data stored in DB,
it is modified to directly extract meta data
that has additional information about media files.
Those information will not stored
in DB and just can be shown when it is requested.
Related API are getAudioMetadata / getVideoMetadata /
getImageMetadata.

2. com.webos.service.mediaindexer.media kind has been
disabled.

3. In taglibextractor, sample rate, bitrate, channel
information extraction functionalities are added. Duration
extraction has been modified.

4. In gstreamerextractor, video and audio codec information
extraction functionalities are added.

:Testing Performed:
Tested with webos Auto

:QA Notes:

:Issues Addressed:
[PLAT-116006] Implement direct meta data extraction

Change-Id: Ibd6b95a7aab638ab23bbc648d76300ddacab17a5
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediaindexer/+/281479
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Jimmy Ohn <yongjin.ohn@lge.com>
Tested-by: Jimmy Ohn <yongjin.ohn@lge.com>