Add check logic whether device is first mounted or not on target
:Release Notes:
Add check logic whether device is first mounted or not on target

:Detailed Notes:
media indexer comapares between scanning media contents in usb and
stored media contents in data base even if usb is first mounted on
target. this compare logic affects performance.
So we need to check logic that device is mounted first on target.

:Testing Performed:
Tested in webOS Auto

:QA Notes:

:Issues Addressed:
[PLAT-128955] Apply device check logic for performance improvement

Change-Id: Id178490f82be6e841f8e1d293dc4db36e82bc582
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediaindexer/+/298652
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: Static Analysis <dtvqe@lge.com>
Reviewed-by: <ejihoon.lee@lge.com>
Tested-by: <ejihoon.lee@lge.com>