Apply video thumbnail extraction in parallel
:Release Notes:
Apply video thumbnail extraction in parallel

:Detailed Notes:
In previous, we had a problem that when use hardware deocder in
parallel, sometimes thumbnail extraction is failed due to hardware
resource problem. To prevent this problem, we added mutex in
thumbnail extraction function. as a result, video thumbnail was
extracted one by one even if threads could extract thumbnail.
But now, we are using software NOT hardware decoder.
So, mutex is removed and resources that could be shared between threads in
GStreamer extractor are moved in thumbnail function to prevent thread
concurrency. And there was some code refactoring.

:Testing Performed:
Tested in webOS Auto

:QA Notes:

:Issues Addressed:
[PLAT-130729] Apply GStreamer thumbnail extraction in parallel

Change-Id: Id7c60b209f899ae05cb6b8d354bc1aaddf2bdb6a
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediaindexer/+/300117
Reviewed-by: Commit Msg Checker <commit_msg@lge.com>
Reviewed-by: Ban Word Checker <ban_word@lge.com>
Reviewed-by: <ejihoon.lee@lge.com>
Tested-by: <ejihoon.lee@lge.com>