Fix race condition occurred in rescan procedure
:Release Notes:
Fix race condition occurred in rescan procedure

:Detailed Notes:
In rescan procedure, when target device is booted after
injecting already known usb, race condition can be
occurred because the service recognize the usb as new device
for a couple of seconds. By modifying injectKnownDevices to operate
as atomically, this issue can be handled.

:Testing Performed:
Tested in webOS Auto

:QA Notes:

:Issues Addressed:
[PLAT-130851] [Performance Improvement] Fix race condition
              in rescan process

Change-Id: I6b8148c57952f9a517628fd3ffa720555d744fca
Reviewed-on: http://gpro.lge.com/c/webosose/com.webos.service.mediaindexer/+/300640
Tested-by: <ejihoon.lee@lge.com>
Reviewed-by: <ejihoon.lee@lge.com>